backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2